Task5_Real-time Multi-Object Colorization with Semantic Segmentation/requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
streamlit-webrtc==0.47.1
av==10.0.0

